PageBranch widget for Dashboard (#848)
Removed the use of IndexedStack widget for retrieving page screens in dashboard
Isolated and modified PageBranch widget to be used for both mobile and
desktop dashboard
