+++ lib/screens/common_widgets/common_widgets.dart
@@ -16,3 +16,4 @@ export 'envvar_span.dart';
 export 'sidebar_filter.dart';
 export 'sidebar_header.dart';
 export 'sidebar_save_button.dart';
+export 'dashboard_page_branch.dart';
