+++ lib/screens/mobile/dashboard.dart
@@ -4,12 +4,8 @@ import 'package:flutter/services.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flex_color_scheme/flex_color_scheme.dart';
 import 'package:apidash/providers/providers.dart';
-import 'requests_page/requests_page.dart';
-import '../envvar/environment_page.dart';
-import '../history/history_page.dart';
-import '../settings_page.dart';
-import 'widgets/page_base.dart';
 import 'navbar.dart';
+import '../common_widgets/common_widgets.dart';
 
 class MobileDashboard extends ConsumerStatefulWidget {
   const MobileDashboard({super.key});
@@ -55,30 +51,5 @@ class _MobileDashboardState extends ConsumerState<MobileDashboard> {
         ],
       ),
     );
-  }
-}
-
-class PageBranch extends ConsumerWidget {
-  const PageBranch({
-    super.key,
-    required this.pageIndex,
-  });
-
-  final int pageIndex;
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    switch (pageIndex) {
-      case 1:
-        return const EnvironmentPage();
-      case 2:
-        return const HistoryPage();
-      case 3:
-        return const PageBase(
-          title: 'Settings',
-          scaffoldBody: SettingsPage(),
-        );
-      default:
-        return const RequestResponsePage();
-    }
   }
 }
